Add and refactor Product repository tests
