--- src/test/java/id/ac/ui/cs/advprog/eshop/repository/DeleteProductTest.java
@@ -1,67 +1,0 @@
-package id.ac.ui.cs.advprog.eshop.repository;
-
-import id.ac.ui.cs.advprog.eshop.model.Product;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-import java.util.Iterator;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-@ExtendWith(MockitoExtension.class)
-class DeleteProductTest {
-    @InjectMocks
-    ProductRepository productRepository;
-
-    @BeforeEach
-    void setUp() {
-    }
-
-    @Test
-    void testDeleteAndFind() {
-        Product product = new Product();
-        product.setProductId("eb558e9f-1c39-460e-8860-71af6af63bd6");
-        product.setProductName("Sampo Cap Bambang");
-        product.setProductQuantity(100);
-        productRepository.create(product);
-
-        productRepository.delete(product.getProductId());
-
-        Iterator<Product> productIterator = productRepository.findAll();
-        assertFalse(productIterator.hasNext());
-    }
-
-    @Test
-    void testDeleteIfMoreThanOneProduct() {
-        Product product1 = new Product();
-        product1.setProductId("eb558e9f-1c39-460e-8860-71af6af63bd6");
-        product1.setProductName("Sampo Cap Bambang");
-        product1.setProductQuantity(100);
-        productRepository.create(product1);
-
-        Product product2 = new Product();
-        product2.setProductId("a0f9de46-90b1-437d-a0bf-d0b21dde9096");
-        product2.setProductName("Sampo Cap Aris");
-        product2.setProductQuantity(50);
-        productRepository.create(product2);
-
-        Product product3 = new Product();
-        product3.setProductId("59e1c9f6-ae90-4313-b57b-70604b705f55");
-        product3.setProductName("Sampo Cap Kunta");
-        product3.setProductQuantity(20);
-        productRepository.create(product3);
-
-        productRepository.delete(product2.getProductId());
-
-        Iterator<Product> productIterator = productRepository.findAll();
-        assertTrue(productIterator.hasNext());
-        Product savedProduct = productIterator.next();
-        assertEquals(product1.getProductId(), savedProduct.getProductId());
-        savedProduct = productIterator.next();
-        assertEquals(product3.getProductId(), savedProduct.getProductId());
-        assertFalse(productIterator.hasNext());
-    }
-}
