--- src/test/java/id/ac/ui/cs/advprog/eshop/repository/EditProductTest.java
@@ -1,44 +1,0 @@
-package id.ac.ui.cs.advprog.eshop.repository;
-
-import id.ac.ui.cs.advprog.eshop.model.Product;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.InjectMocks;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-import java.util.Iterator;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-@ExtendWith(MockitoExtension.class)
-class EditProductTest {
-    @InjectMocks
-    ProductRepository productRepository;
-
-    @BeforeEach
-    void setUp() {
-    }
-
-    @Test
-    void testEditAndFind() {
-        Product product = new Product();
-        product.setProductId("eb558e9f-1c39-460e-8860-71af6af63bd6");
-        product.setProductName("Sampo Cap Bambang");
-        product.setProductQuantity(100);
-        productRepository.create(product);
-
-        Product editDetails = new Product();
-        editDetails.setProductName("Sampo Cap Bimo");
-        editDetails.setProductQuantity(50);
-        productRepository.edit(product.getProductId(), editDetails);
-
-        Iterator<Product> productIterator = productRepository.findAll();
-        assertTrue(productIterator.hasNext());
-        Product savedProduct = productIterator.next();
-        assertEquals(product.getProductId(), savedProduct.getProductId());
-        assertEquals("Sampo Cap Bimo", savedProduct.getProductName());
-        assertEquals(50, savedProduct.getProductQuantity());
-        assertFalse(productIterator.hasNext());
-    }
-}
